Removed plugin, added composables
