+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@sidebase/nuxt-pdf",
-  "version": "0.1.1-alpha.13",
+  "version": "0.1.1-alpha.15",
   "license": "MIT",
   "type": "module",
   "exports": {
